Move Options and Validation to package
