+++ pkg/apis/logging/realclientip.go
@@ -1,19 +1,18 @@
-package main
+package logging
 
 import (
 	"fmt"
 	"net"
 	"net/http"
 	"strings"
-
-	"github.com/oauth2-proxy/oauth2-proxy/pkg/logger"
 )
 
-type realClientIPParser interface {
+// RealClientIPParser is an interface for a getting the client's real IP to be used for logging.
+type RealClientIPParser interface {
 	GetRealClientIP(http.Header) (net.IP, error)
 }
 
-func getRealClientIPParser(headerKey string) (realClientIPParser, error) {
+func GetRealClientIPParser(headerKey string) (RealClientIPParser, error) {
 	headerKey = http.CanonicalHeaderKey(headerKey)
 
 	switch headerKey {
@@ -73,13 +72,11 @@ func getRemoteIP(req *http.Request) (net.IP, error) {
 	}
 }
 
-// getClientString obtains the human readable string of the remote IP and optionally the real client IP if available
-func getClientString(p realClientIPParser, req *http.Request, full bool) (s string) {
+// GetClientString obtains the human readable string of the remote IP and optionally the real client IP if available
+func GetClientString(p RealClientIPParser, req *http.Request, full bool) (s string) {
 	var realClientIPStr string
 	if p != nil {
-		if realClientIP, err := p.GetRealClientIP(req.Header); err != nil {
-			logger.Printf("Unable to get real client IP: %v", err)
-		} else if realClientIP != nil {
+		if realClientIP, err := p.GetRealClientIP(req.Header); err == nil && realClientIP != nil {
 			realClientIPStr = realClientIP.String()
 		}
 	}
@@ -87,9 +84,6 @@ func getClientString(p realClientIPParser, req *http.Request, full bool) (s stri
 	var remoteIPStr string
 	if remoteIP, err := getRemoteIP(req); err == nil {
 		remoteIPStr = remoteIP.String()
-	} else {
-		// Should not happen, if it does, likely a bug.
-		logger.Printf("Unable to get remote IP(?!?!): %v", err)
 	}
 
 	if !full && realClientIPStr != "" {
